+++ js/repositories/directoryRepository.js
@@ -1,0 +1,6 @@
+"use strict";
+
+const DirectoryRepository = createRepository(
+    db.directory,
+    () => db.directory.orderBy("updatedAt").reverse().toArray()
+);